requirements.txt
mediapipe==0.10.2
numpy==1.25.1
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
pickle5==0.0.11
pickles==0.1.1
Pillow==9.5.0
streamlit==1.24.1
scikit-learn==1.1.2
streamlit-webrtc